[HienHT] [AUTH_02] Fix authentication + remove firebase credentials

+++ Infrastructure/Contexts/BookingActivity.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.Contexts
         public int Id { get; set; }
         public string Description { get; set; }
         public DateTime CreatedAt { get; set; }
-        public long BookingId { get; set; }
+        public int BookingId { get; set; }
 
         public virtual Booking Booking { get; set; }
     }
