+++ Infrastructure/Contexts/FeedBack.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.Contexts
         public int Id { get; set; }
         public int ImageUrl { get; set; }
         public int RateScore { get; set; }
-        public long BookingId { get; set; }
+        public int BookingId { get; set; }
 
         public virtual Booking Booking { get; set; }
     }
